feature/ add container & component file
