+++ src/container/RoomList.jsx
@@ -1,0 +1,20 @@
+import { useState, useEffect } from "react";
+import  RoomItems  from '../component/RoomItems';
+
+
+function RoomList() {
+  return (
+    <>
+      <ul className="flex flex-wrap justify-end">
+        <RoomItems/>
+        <RoomItems/>
+        <RoomItems/>
+        <RoomItems/>
+        <RoomItems/>
+        <RoomItems/>
+      </ul>
+    </>
+  )
+  }
+  
+  export default RoomList;
