--- src/pages/container/RoomList.jsx
@@ -1,13 +1,0 @@
-import { useState, useEffect } from "react";
-import RoomItems from './RoomItems';
-
-
-function RoomList() {
-  return (
-    <>
-    <RoomItems/>
-    </>
-  );
-  }
-  
-  export default RoomList;
