--- src/pages/container/component/RoomItems.jsx
@@ -1,12 +1,0 @@
-import { useState, useEffect } from "react";
-
-function RoomItems() {
-  return (
-    <>
-    console.log("RoomList");
-    </>
-
-  );
-  }
-  
-export default RoomItems;
